+++ src/lib/component/SettingDialog/reflectImmediately/bindChangeLockConfig.js
@@ -1,5 +1,5 @@
 import delgate from 'delegate'
-import debounce300 from '../../debounce300'
+import debounce300 from './debounce300'
 
 export default function (content, typeDictionary) {
   delgate(
